FE API: all images are now batch loaded
- the data loading logic was moved from project base to the frontend-api package
