+++ packages/framework/src/Component/Image/Image.php
@@ -288,4 +288,19 @@ class Image extends AbstractTranslatableEntity implements EntityFileUploadInterf
             );
         }
     }
+
+    /**
+     * @param string|null $friendlyUrlSlug
+     * @return string
+     */
+    public function getSeoFilename(?string $friendlyUrlSlug): string
+    {
+        $slug = '';
+
+        if ($friendlyUrlSlug !== null) {
+            $slug = $friendlyUrlSlug . '_';
+        }
+
+        return  $slug . $this->id . '.' . $this->extension;
+    }
 }
