+++ packages/frontend-api/src/Component/Image/ImagesBatchLoader.php
@@ -2,36 +2,35 @@
 
 declare(strict_types=1);
 
-namespace App\FrontendApi\Model\Image;
+namespace Shopsys\FrontendApiBundle\Component\Image;
 
-use App\Component\Image\Image;
-use App\Component\Image\ImageFacade;
-use App\FrontendApi\Model\Image\ImageFacade as FrontendApiImageFacade;
 use GraphQL\Executor\Promise\Promise;
 use GraphQL\Executor\Promise\PromiseAdapter;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
 use Shopsys\FrameworkBundle\Component\Image\Config\ImageEntityConfig;
 use Shopsys\FrameworkBundle\Component\Image\Exception\ImageNotFoundException;
+use Shopsys\FrameworkBundle\Component\Image\Image;
+use Shopsys\FrameworkBundle\Component\Image\ImageFacade;
 use Shopsys\FrameworkBundle\Component\Utils\Utils;
 
 class ImagesBatchLoader
 {
     /**
      * @param \GraphQL\Executor\Promise\PromiseAdapter $promiseAdapter
-     * @param \App\FrontendApi\Model\Image\ImageFacade $frontendApiImageFacade
-     * @param \App\Component\Image\ImageFacade $imageFacade
+     * @param \Shopsys\FrontendApiBundle\Component\Image\ImageApiFacade $imageApiFacade
+     * @param \Shopsys\FrameworkBundle\Component\Image\ImageFacade $imageFacade
      * @param \Shopsys\FrameworkBundle\Component\Domain\Domain $domain
      */
     public function __construct(
-        private PromiseAdapter $promiseAdapter,
-        private FrontendApiImageFacade $frontendApiImageFacade,
-        private ImageFacade $imageFacade,
-        private Domain $domain,
+        protected readonly PromiseAdapter $promiseAdapter,
+        protected readonly ImageApiFacade $imageApiFacade,
+        protected readonly ImageFacade $imageFacade,
+        protected readonly Domain $domain,
     ) {
     }
 
     /**
-     * @param \App\FrontendApi\Model\Image\ImageBatchLoadData[] $imagesBatchLoadData
+     * @param \Shopsys\FrontendApiBundle\Component\Image\ImageBatchLoadData[] $imagesBatchLoadData
      * @return \GraphQL\Executor\Promise\Promise
      */
     public function loadByBatchData(array $imagesBatchLoadData): Promise
@@ -50,12 +49,12 @@ class ImagesBatchLoader
     }
 
     /**
-     * @param \App\FrontendApi\Model\Image\ImageBatchLoadData[] $imagesBatchLoadData
+     * @param \Shopsys\FrontendApiBundle\Component\Image\ImageBatchLoadData[] $imagesBatchLoadData
      * @param string $entityName
      * @param string $type
      * @return array<string, array|null>
      */
-    private function getImagesByEntityNameAndTypeIndexedByDataId(
+    protected function getImagesByEntityNameAndTypeIndexedByDataId(
         array $imagesBatchLoadData,
         string $entityName,
         string $type,
@@ -64,7 +63,7 @@ class ImagesBatchLoader
             $type = null;
         }
         $entityIds = array_map(fn (ImageBatchLoadData $imageBatchLoadData) => $imageBatchLoadData->getEntityId(), $imagesBatchLoadData);
-        $imagesIndexedByEntityId = $this->frontendApiImageFacade->getAllImagesIndexedByEntityId($entityIds, $entityName, $type);
+        $imagesIndexedByEntityId = $this->imageApiFacade->getAllImagesIndexedByEntityId($entityIds, $entityName, $type);
 
         $images = [];
 
@@ -82,10 +81,10 @@ class ImagesBatchLoader
     }
 
     /**
-     * @param \App\FrontendApi\Model\Image\ImageBatchLoadData[] $imagesBatchLoadData
-     * @return \App\FrontendApi\Model\Image\ImageBatchLoadData[][][]
+     * @param \Shopsys\FrontendApiBundle\Component\Image\ImageBatchLoadData[] $imagesBatchLoadData
+     * @return \Shopsys\FrontendApiBundle\Component\Image\ImageBatchLoadData[][][]
      */
-    private function getImageBatchLoadDataArrayByEntityAndType(array $imagesBatchLoadData): array
+    protected function getImageBatchLoadDataArrayByEntityAndType(array $imagesBatchLoadData): array
     {
         $result = [];
 
@@ -100,10 +99,10 @@ class ImagesBatchLoader
 
     /**
      * @param array<string, array|null> $allImagesIndexedByImageBatchLoadDataId
-     * @param \App\FrontendApi\Model\Image\ImageBatchLoadData[] $imagesBatchLoadData
+     * @param \Shopsys\FrontendApiBundle\Component\Image\ImageBatchLoadData[] $imagesBatchLoadData
      * @return array<int, array|null>
      */
-    private function sortAllImagesByOriginalInputData(
+    protected function sortAllImagesByOriginalInputData(
         array $allImagesIndexedByImageBatchLoadDataId,
         array $imagesBatchLoadData,
     ): array {
@@ -122,10 +121,10 @@ class ImagesBatchLoader
     }
 
     /**
-     * @param \App\Component\Image\Image[] $images
+     * @param \Shopsys\FrameworkBundle\Component\Image\Image[] $images
      * @return array<int, array{url: string, name: string|null}>
      */
-    private function getResolvedImages(array $images): array
+    protected function getResolvedImages(array $images): array
     {
         $resolvedImages = [];
 
@@ -141,10 +140,10 @@ class ImagesBatchLoader
     }
 
     /**
-     * @param \App\Component\Image\Image $image
+     * @param \Shopsys\FrameworkBundle\Component\Image\Image $image
      * @return array{url: string, name: string|null}
      */
-    private function getResolvedImage(Image $image): array
+    protected function getResolvedImage(Image $image): array
     {
         return [
             'url' => $this->imageFacade->getImageUrl(
