+++ packages/frontend-api/src/Model/Resolver/Image/ImagesQuery.php
@@ -4,69 +4,61 @@ declare(strict_types=1);
 
 namespace Shopsys\FrontendApiBundle\Model\Resolver\Image;
 
+use GraphQL\Executor\Promise\Promise;
+use Overblog\DataLoader\DataLoaderInterface;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
 use Shopsys\FrameworkBundle\Component\Image\Config\ImageConfig;
-use Shopsys\FrameworkBundle\Component\Image\Image;
 use Shopsys\FrameworkBundle\Component\Image\ImageFacade;
-use Shopsys\FrameworkBundle\Model\Advert\Advert;
-use Shopsys\FrameworkBundle\Model\Product\Product;
-use Shopsys\FrontendApiBundle\Component\Image\ImageFacade as FrontendApiImageFacade;
+use Shopsys\FrontendApiBundle\Component\Image\ImageApiFacade;
+use Shopsys\FrontendApiBundle\Component\Image\ImageBatchLoadData;
 use Shopsys\FrontendApiBundle\Model\Resolver\AbstractQuery;
 
 class ImagesQuery extends AbstractQuery
 {
-    protected const IMAGE_ENTITY_PRODUCT = 'product';
-
     /**
      * @param \Shopsys\FrameworkBundle\Component\Image\ImageFacade $imageFacade
      * @param \Shopsys\FrameworkBundle\Component\Image\Config\ImageConfig $imageConfig
      * @param \Shopsys\FrameworkBundle\Component\Domain\Domain $domain
-     * @param \Shopsys\FrontendApiBundle\Component\Image\ImageFacade $frontendApiImageFacade
+     * @param \Shopsys\FrontendApiBundle\Component\Image\ImageApiFacade $imageApiFacade
+     * @param \Overblog\DataLoader\DataLoaderInterface $imagesBatchLoader
+     * @param \Overblog\DataLoader\DataLoaderInterface $firstImageBatchLoader
      */
     public function __construct(
         protected readonly ImageFacade $imageFacade,
         protected readonly ImageConfig $imageConfig,
         protected readonly Domain $domain,
-        protected readonly FrontendApiImageFacade $frontendApiImageFacade,
+        protected readonly ImageApiFacade $imageApiFacade,
+        protected readonly DataLoaderInterface $imagesBatchLoader,
+        protected readonly DataLoaderInterface $firstImageBatchLoader,
     ) {
     }
 
     /**
      * @param object $entity
      * @param string|null $type
-     * @return array
+     * @return \GraphQL\Executor\Promise\Promise
      */
-    public function imagesByEntityQuery(object $entity, ?string $type): array
+    public function mainImageByEntityPromiseQuery(object $entity, ?string $type): Promise
     {
-        $entityName = $this->imageConfig->getEntityName($entity);
+        $imageEntityConfig = $this->imageConfig->getImageEntityConfig($entity);
 
-        return $this->resolveByEntityId($entity->getId(), $entityName, $type);
+        return $this->mainImageByEntityIdPromiseQuery($entity->getId(), $imageEntityConfig->getEntityName(), $type);
     }
 
     /**
-     * @param \Shopsys\FrameworkBundle\Model\Product\Product|array $data
+     * @param int $entityId
+     * @param string $entityName
      * @param string|null $type
-     * @return array
+     * @return \GraphQL\Executor\Promise\Promise
      */
-    public function imagesByProductQuery($data, ?string $type): array
-    {
-        $productId = $data instanceof Product ? $data->getId() : $data['id'];
-
-        return $this->resolveByEntityId($productId, static::IMAGE_ENTITY_PRODUCT, $type);
-    }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Model\Advert\Advert $advert
-     * @param string|null $type
-     * @return array
-     */
-    public function imagesByAdvertQuery(Advert $advert, ?string $type): array
-    {
-        $entityName = $this->imageConfig->getEntityName($advert);
-
-        return $this->getResolvedImages(
-            $this->frontendApiImageFacade->getImagesByEntityIdAndNameIndexedById(
-                $advert->getId(),
+    public function mainImageByEntityIdPromiseQuery(
+        int $entityId,
+        string $entityName,
+        ?string $type,
+    ): Promise {
+        return $this->firstImageBatchLoader->load(
+            new ImageBatchLoadData(
+                $entityId,
                 $entityName,
                 $type,
             ),
@@ -74,46 +66,34 @@ class ImagesQuery extends AbstractQuery
     }
 
     /**
+     * @param object $entity
+     * @param string|null $type
+     * @return \GraphQL\Executor\Promise\Promise
+     */
+    public function imagesByEntityPromiseQuery(object $entity, ?string $type): Promise
+    {
+        $imageEntityConfig = $this->imageConfig->getImageEntityConfig($entity);
+
+        return $this->resolveByEntityIdPromise($entity->getId(), $imageEntityConfig->getEntityName(), $type);
+    }
+
+    /**
      * @param int $entityId
      * @param string $entityName
      * @param string|null $type
-     * @return array
+     * @return \GraphQL\Executor\Promise\Promise
      */
-    protected function resolveByEntityId(int $entityId, string $entityName, ?string $type): array
-    {
-        $images = $this->frontendApiImageFacade->getImagesByEntityIdAndNameIndexedById($entityId, $entityName, $type);
-
-        return $this->getResolvedImages($images);
-    }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Component\Image\Image[] $images
-     * @return array<int, array{url: string, name: string|null}>
-     */
-    protected function getResolvedImages(array $images): array
-    {
-        $resolvedImages = [];
-
-        foreach ($images as $image) {
-            $resolvedImages[] = $this->getResolvedImage($image);
-        }
-
-        return $resolvedImages;
-    }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Component\Image\Image $image
-     * @return array{url: string, name: string|null}
-     */
-    protected function getResolvedImage(Image $image): array
-    {
-        return [
-            'name' => $image->getName(),
-            'url' => $this->imageFacade->getImageUrl(
-                $this->domain->getCurrentDomainConfig(),
-                $image,
-                $image->getType(),
+    protected function resolveByEntityIdPromise(
+        int $entityId,
+        string $entityName,
+        ?string $type,
+    ): Promise {
+        return $this->imagesBatchLoader->load(
+            new ImageBatchLoadData(
+                $entityId,
+                $entityName,
+                $type,
             ),
-        ];
+        );
     }
 }
