+++ project-base/app/src/Component/Image/Image.php
@@ -56,19 +56,4 @@ class Image extends BaseImage
     {
         $this->akeneoImageType = $akeneoImageType;
     }
-
-    /**
-     * @param string|null $friendlyUrlSlug
-     * @return string
-     */
-    public function getSeoFilename(?string $friendlyUrlSlug): string
-    {
-        $slug = '';
-
-        if ($friendlyUrlSlug !== null) {
-            $slug = $friendlyUrlSlug . '_';
-        }
-
-        return  $slug . $this->id . '.' . $this->extension;
-    }
 }
